refactor(city): render filter pills in GlobalMap, under the header
Moves the category filter pills out of the City guide's bottom sheet and
into GlobalMap, rendered as a floating row under GlobalMapHeader. Pills
now call the map's own handleFilterChange directly instead of round-tripping
through the filters:change EventEmitter, since both now live in the same
component.
